recipes fixes (e.g, update readme, google drive, add data-io prep in train.py)
